+++ src/main/java/net/geforcemods/securitycraft/network/server/AssembleBlockPocket.java
@@ -1,8 +1,14 @@
 package net.geforcemods.securitycraft.network.server;
 
+import net.geforcemods.securitycraft.SCContent;
 import net.geforcemods.securitycraft.blockentities.BlockPocketManagerBlockEntity;
+import net.geforcemods.securitycraft.util.PlayerUtils;
+import net.geforcemods.securitycraft.util.Utils;
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.FriendlyByteBuf;
+import net.minecraft.network.chat.MutableComponent;
+import net.minecraft.world.entity.player.Player;
 import net.neoforged.neoforge.network.NetworkEvent;
 
 public class AssembleBlockPocket {
@@ -27,10 +33,17 @@ public class AssembleBlockPocket {
 	}
 
 	public void handle(NetworkEvent.Context ctx) {
-		if (ctx.getSender().level().getBlockEntity(pos) instanceof BlockPocketManagerBlockEntity be && be.isOwnedBy(ctx.getSender())) {
+		Player player = ctx.getSender();
+
+		if (player.level().getBlockEntity(pos) instanceof BlockPocketManagerBlockEntity be && be.isOwnedBy(player)) {
+			MutableComponent feedback;
+
 			be.setSize(size);
-			be.autoAssembleMultiblock();
+			feedback = be.autoAssembleMultiblock();
 			be.setChanged();
+
+			if (feedback != null)
+				PlayerUtils.sendMessageToPlayer(player, Utils.localize(SCContent.BLOCK_POCKET_MANAGER.get().getDescriptionId()), feedback, ChatFormatting.DARK_AQUA);
 		}
 	}
 }